Fix for snakemake's knit_root_dir parameter passed to rmarkdown::render
It fixes an issue with rmarkdown crashing during QC reports rules of the WGBS pipeline. Before this patch, rmarkdown 0.9.3 is installed by default, which doesn't take into account snakemake's knit_root_dir parameter passed to rmarkdown::render.

This patch should fix it (I also encountered errors about htmltools and yaml packages that are ̀installed by an R version with different internals; it needs to be reinstalled for use with this R version`, I'm not sure this patch fixes that).